# Model SyntaxVisitor as a class