… Sources/SwiftSyntax/SyntaxChildren.swift
… Sources/SwiftSyntax/SyntaxCollections.swift.gyb
… Sources/SwiftSyntax/SyntaxFactory.swift.gyb
… Sources/SwiftSyntax/SyntaxKind.swift.gyb
… Sources/SwiftSyntax/SyntaxVerifier.swift
… Sources/SwiftSyntaxBuilder/DeclBuildables.swift
… Sources/SwiftSyntaxBuilder/ExprBuildables.swift
… Sources/SwiftSyntaxBuilder/SyntaxBuildables.swift
… Sources/lit-test-helper/main.swift
… Tests/SwiftSyntaxTest/AbsolutePosition.swift
… Tests/SwiftSyntaxTest/ClassificationTests.swift
… Tests/SwiftSyntaxTest/CustomReflecatbleTests.swift
… Tests/SwiftSyntaxTest/DiagnosticTest.swift
… Tests/SwiftSyntaxTest/IncrementalParsingTests.swift
… Tests/SwiftSyntaxTest/ModifierTests.swift
… Tests/SwiftSyntaxTest/ParseFile.swift
… Tests/SwiftSyntaxTest/SyntaxChildren.swift
… Tests/SwiftSyntaxTest/SyntaxCollections.swift
… Tests/SwiftSyntaxTest/SyntaxFactory.swift
… Tests/SwiftSyntaxTest/VisitorTest.swift