Since all the methods on the SyntaxVisitor were mutating, it makes more sense to model the SyntaxVisitor as a class. This also increases visitation performance.

## API Change

### `SyntaxVisitor`

* Structs that conformed to `SyntaxVisitor` need to become classes
* Otherwise, the change should be source compatible

### `SyntaxAnyVisitor`

* Structs that conformed to `SyntaxAnyVisitor` need to become classes
* Since `SyntaxAnyVisitor` no longer specializes `SyntaxVisitor` the methods were renamed from `visitAny` to the more generic name `visit(_ syntax: Syntax)`. The corresponding renaming needs to be done in the codebase.

## Performance improvements

Performance measurements were done by running an empty `SyntaxRewriter` / `SyntaxVisitor` on a Swift file with ~5000 lines of code. The baseline is the current performance of SwiftSyntax before #155 is merged.

### Empty `SyntaxVisitor`

|               | Before | After #155      | After this PR    |
|---------|--------|----------------|----------------|
| Debug   | 92.9ms | 112.1ms (+21%) | 107.9ms (+16%) |
| Release | 12.3ms | 12.6ms (+3%)   | 10.9ms (-12%)  |

### Empty `SyntaxAnyVisitor`

|         | Before  | After #155          | After this PR   |
|---------|---------|----------------|---------------|
| Debug   | 106.2ms | 136.2ms (+28%) | 115.3 (+9%) |
| Release | 24.8ms  | 15.5ms (-38%)  | 11.4ms (-54%)  |